[Header Bar Logo] Replace the PNG logo assets to SVG
Replace the PNG logo assets to SVG in order to maintain pixel-perfect quality.

Add media breakpoint to switch the long version of the logo to a short one to fit mobile resolutions.

Note: Might need to optimize the SVGs later since their file sizes are quite larger than the PNGs' for some reason.